% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{DBSCAN}
\alias{DBSCAN}
\title{DBSCAN}
\usage{
DBSCAN(vec, eps, min_samples)
}
\arguments{
\item{vec}{텍스트 벡터}

\item{eps}{epsilon 값(보폭)}

\item{min_samples}{군집을 형성하는 최소 샘플 갯수}
}
\description{
DBSCAN 알고리즘을 진행합니다.
}
\examples{
df <- data.frame(
  "일자" = c(20230101, 20230102, 20230103),
  "언론사" = c("조선일보", "한겨례", "경향신문"),
  "제목" = c("[속보] 경찰, 민주노총 도심 집회 1차 해산명령", 
              "[사설] 한반도 긴장 높인 북한의 군사정찰위성 발사 규탄한다",
              "‘직권남용죄’ 남용 유감"),
  "키워드" = c("경찰,해산명령,민주,노총,도심,집회,해산,명령,민주노총",
                "한반도,긴장,북한,규탄,군사,정찰,위성,발사,북한,인공위성",
                "직권남용죄,남용,유감,왜냐면,이정환,민주사회,변호사모임"))

vec <- tfidf_vector(df)
cluster <- DBSCAN(vec, eps = 0.1, min_samples = 2)

}
