% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{duplication_remover}
\alias{duplication_remover}
\title{중복 값 제거}
\usage{
duplication_remover(news_key)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
중복 값 제거
}
