% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{keyword_dataframe}
\alias{keyword_dataframe}
\title{df to Keyword_dataframe}
\usage{
keyword_dataframe(df)
}
\arguments{
\item{df}{BigKinds 원본 문서}
}
\value{
키워드 데이터프레임으로 변환
}
\description{
df to Keyword_dataframe
}
