% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{keywords_wordcloud}
\alias{keywords_wordcloud}
\title{keywords_wordcloud}
\usage{
keywords_wordcloud(df, press = NA)
}
\arguments{
\item{df}{BigKinds 원본 문서}

\item{press}{확인할 언론사 이름}
}
\description{
언론사 별로 가장 많이 등장한 단어 순위를 wordcloud로 시각화합니다.
}
\examples{
keywords_wordcloud(df, "조선일보")
}
