% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{network_graph}
\alias{network_graph}
\title{word network}
\usage{
network_graph(dcm, topwords = 50, min_freq = 0.5)
}
\arguments{
\item{dcm}{단어 동시출현행렬}

\item{topwords}{feature 단어 갯수, 최대 50개까지 가능}

\item{min_freq}{그래프를 형성할 연결 최소 빈도 수}
}
\description{
단어 네트워크 분석 결과를 시각화합니다.
}
