% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBigKinds.R
\docType{package}
\name{RBigKinds-package}
\alias{RBigKinds}
\alias{RBigKinds-package}
\title{RBigKinds: BigKinds Data Analysis Toolkit for R}
\description{
RBigKinds는 한국 언론의 빅데이터 저장소인 BigKinds에서 추출한 데이터를
low-Code로 분석할 수 있게 만든 툴입니다.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sorrychoe.github.io/RBigKinds/}
  \item \url{https://github.com/sorrychoe/RBigKinds}
  \item Report bugs at \url{https://github.com/sorrychoe/RBigKinds/issues}
}

}
\author{
\strong{Maintainer}: Jaeseong Choe \email{cjssoote@gmail.com}

Authors:
\itemize{
  \item Jaeseong Choe \email{cjssoote@gmail.com}
}

}
\keyword{internal}
