% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{top_words}
\alias{top_words}
\title{top_words}
\usage{
top_words(df, press = NA, top_n = 25)
}
\arguments{
\item{df}{BigKinds 원본 문서}

\item{press}{확인할 언론사 이름}

\item{top_n}{시각화할 단어 갯수}
}
\description{
언론사 별로 가장 많이 등장한 단어 순위를 시각화합니다.
최대 몇개의 단어를 추출할지는 직접 정할 수 있습니다.
default는 25개입니다.
}
