% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{association}
\alias{association}
\title{association}
\usage{
association(df, min_support = 0.5, minlen = 2, maxlen = 10)
}
\arguments{
\item{df}{BigKinds 원본 문서}

\item{min_support}{최소 지지도}

\item{minlen}{연관된 최소 갯수}

\item{maxlen}{연관된 최대 갯수}
}
\description{
기사에 등장한 단어 별로 연관분석을 진행합니다.
연관분석 방법은 Apriori입니다.
}
\examples{
association(df, min_support = 0.6, minlen = 3, maxlen = 10)
}
