% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{counter_to_dataframe}
\alias{counter_to_dataframe}
\title{counter dict를 dataframe으로 변환}
\usage{
counter_to_dataframe(key_words)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
counter dict를 dataframe으로 변환
}
