% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{header_remover}
\alias{header_remover}
\title{header_remover}
\usage{
header_remover(df)
}
\arguments{
\item{df}{BigKinds 원본 문서}
}
\description{
상단에 존재하는 헤더를 제거합니다.
}
\examples{
data <- header_remover(df)
head(data)
}
