% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{Kmeans}
\alias{Kmeans}
\title{Kmeans}
\usage{
Kmeans(vec, k, random_state = 123)
}
\arguments{
\item{k}{형성할 군집 갯수}

\item{random_state}{seed 값}

\item{df}{BigKinds 원본 문서}
}
\description{
kmeans clustering을 진행합니다.
}
\examples{
Kmeans(df, k = 3)

}
