% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{Kmeans}
\alias{Kmeans}
\title{Kmeans}
\usage{
Kmeans(vec, k, max = 1000, random_state = 123)
}
\arguments{
\item{vec}{텍스트 벡터}

\item{k}{형성할 군집 갯수}

\item{max}{최대 반복 횟수}

\item{random_state}{seed 값}
}
\description{
kmeans clustering을 진행합니다.
}
\examples{
Kmeans(vec, k = 3, max = 5000)

}
