% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{topic_modeling}
\alias{topic_modeling}
\title{topic_modeling}
\usage{
topic_modeling(df, k, method = "Gibbs")
}
\arguments{
\item{df}{BigKinds 원본 문서}

\item{k}{토픽 개수}
}
\description{
토픽 모델링을 시행합니다.
}
\examples{
topic_modeling(df, k = 10)

}
