% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{word_tokenizer}
\alias{word_tokenizer}
\title{word_tokenizer}
\usage{
word_tokenizer(df)
}
\arguments{
\item{df}{BigKinds 원본 문서}
}
\description{
파일로부터 문서 별 키워드로 나열된 데이터 프레임으로 변환합니다.
}
\examples{
data <- word_tokenizer(df)
view(data)
}
